updated models.py

+++ models.py
@@ -504,7 +504,7 @@ class DECSeq5(torch.nn.Module):
                 if i+self.k/2 > eidx[0,-1]:
                     e2 = torch.cat([e2,torch.cat([torch.arange(i-(self.k-(eidx[0,-1]-i)),i),torch.arange(i+1,eidx[0,-1]+1)],dim=0)])
                 else:
-                    e2 = torch.cat([e2,torch.cat([torch.arange(i-self.k/2,i),torch.arange(i+1,i+self.k/2+1)],dim=0)])
+                    e2 = torch.cat([e2,torch.cat([torch.arange(i-self.k/2,i),torch.arange(i+1,i+self.k/2+1)],dtype=torch.long,dim=0)])
             if i==eidx[0,int(eidx.shape[1]/2)-1]:
                 e2 = torch.cat([e2,torch.cat([torch.arange(i-1,i-self.k,-1),torch.tensor([i+1])])])
         e2 = torch.cat([e2,torch.arange(eidx[0,-1]-1,(eidx[0,-1]-1)-self.k, -1)],dim=0)
